// 监控/运维大屏/grafana-v11.6.0/public/app/extensions/query-library/types.ts
import { DataQuery, DataSourceRef } from '@grafana/schema';

import { User } from './api/types';

export type QueryTemplateRow = {
  index: string;
  datasourceName?: string;
  description?: string;
  query: DataQuery;
  queryText?: string;
  datasourceRef?: DataSourceRef | null;
  datasourceType?: string;
  createdAtTimestamp?: number;
  user?: User;
  uid?: string;
};